settings: 	* add fetch limits options
libgsql:
	*sqleditor (fetching)
	*bug fixes
	*session (rewrited mutex routines. GMutex)
	
mysql-engine:
	*bug fix (case MYSQL_TYPE_NEWDECIMAL, case MYSQL_TYPE_DECIMAL)
	
other:
	*add engine_oracle.ui file
	*new pictures for the fetching buttons



git-svn-id: http://gsql.googlecode.com/svn/trunk@4 9d50599d-b153-0410-81bd-2571f06fed16

+++ libgsql/cursor.h
@@ -81,7 +81,7 @@ struct _GSQLCursor
 	gboolean	scrollable;
 	
 	GSQLCursorStmtType stmt_type;
-	gdouble		stmt_affected_rows;
+	guint64		stmt_affected_rows;
 	
 	GSQLCursorPrivate *private;
 };
